admin-web/requirements.txt
asgiref==3.7.2
Django==4.2.2
django-split-settings==1.2.0
psycopg2==2.9.6
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
